robyn_training/requirements.txt
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Robyn Python API (Primary approach - Pure Python!)
# Official Python implementation of Meta's Robyn MMM
# https://facebookexperimental.github.io/Robyn/docs/robyn-api
robynpy

# Optional: R integration for alternative approach
# Uncomment if using train_model.R or train_model.py (rpy2 version)
# rpy2>=3.5.0
